+++ volmdlr/discrete_representation.py
@@ -1409,16 +1409,16 @@ class MatrixBasedVoxelization(Voxelization):
         if self.voxel_size != other.voxel_size:
             raise ValueError("Voxel sizes must be the same to perform boolean operations.")
 
-        self_min, self_max = np.array(self.min_grid_center), np.array(self.min_grid_center) + 1
+        self_min, self_max = np.array(self.min_grid_center), np.array(self.max_grid_center) + 1
         other_min, other_max = (
             np.array(other.min_grid_center),
-            np.array(other.min_grid_center) + 1,
+            np.array(other.max_grid_center) + 1,
         )
 
         global_min = np.min([self_min, other_min], axis=0)
         global_max = np.max([self_max, other_max], axis=0)
 
-        new_shape = np.round((global_max - global_min) / self.voxel_size, DECIMALS).astype(int) + 2
+        new_shape = np.round((global_max - global_min) / self.voxel_size, DECIMALS).astype(int)
 
         new_self = np.zeros(new_shape, dtype=np.bool_)
         new_other = np.zeros(new_shape, dtype=np.bool_)
@@ -2376,7 +2376,7 @@ class MatrixBasedPixelization(Pixelization):
         global_min = np.min([self_min, other_min], axis=0)
         global_max = np.max([self_max, other_max], axis=0)
 
-        new_shape = np.round((global_max - global_min) / self.pixel_size, DECIMALS).astype(int) + 2
+        new_shape = np.round((global_max - global_min) / self.pixel_size, DECIMALS).astype(int)
 
         new_self = np.zeros(new_shape, dtype=np.bool_)
         new_other = np.zeros(new_shape, dtype=np.bool_)